Version 1.6
Corrected the way the end of the current class was determined (previously, sometimes the inserted methods would be inserted outside of the class body).  Also, I changed the test to see if the script was previously loaded as it seems that the script was not loaded as the variable b:did_ftpplugin is set by lots of scripts.